+++ ci/install-miniconda.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+# To return a failure if any commands inside fail
+set -e
+
 MINICONDA_URL="http://repo.continuum.io/miniconda"
 
 if [ "$TRAVIS_OS_NAME" == "osx" ]; then
@@ -12,3 +15,7 @@ fi
 wget $MINICONDA_URL/$MINICONDA_FILE -O miniconda.sh
 chmod +x miniconda.sh
 ./miniconda.sh -b
+
+# Workaround for https://github.com/travis-ci/travis-ci/issues/6522
+# Turn off exit on failure.
+set +e